+++ cloveretl.component/src/org/jetel/component/ExtDataGenerator.java
@@ -61,7 +61,7 @@ public class ExtDataGenerator extends DataGenerator {
 	private String generatorClassName;
 	private String generatorURL;
 	private String charset;
-	private int recordsNumber;
+	private long recordsNumber;
 
 	// data generator
 	private Properties generateParameters;
@@ -73,7 +73,7 @@ public class ExtDataGenerator extends DataGenerator {
 	 * @param generate
 	 * @param recordsNumber
 	 */
-	public ExtDataGenerator(String id, RecordGenerate generate, int recordsNumber) {
+	public ExtDataGenerator(String id, RecordGenerate generate, long recordsNumber) {
 		super(id);
 		this.generatorClass = generate;
 		this.recordsNumber = recordsNumber;
@@ -85,7 +85,7 @@ public class ExtDataGenerator extends DataGenerator {
 	 * @param recordsNumber
 	 */
 	public ExtDataGenerator(String id, String generate, String generateClass,
-			String generateURL, int recordsNumber) {
+			String generateURL, long recordsNumber) {
 		super(id);
 		this.generatorSource = generate;
 		this.generatorClassName = generateClass;
@@ -152,7 +152,7 @@ public class ExtDataGenerator extends DataGenerator {
 	 * @throws Exception
 	 */
 	private void executeGenerate(DataRecord[] outRecord) throws Exception {
-		for (int i = 0; (recordsNumber < 0 || i < recordsNumber) && runIt; i++) {
+		for (long i = 0; (recordsNumber < 0 || i < recordsNumber) && runIt; i++) {
 			for (DataRecord oRecord : outRecord)
 				oRecord.reset();
 			int transformResult = -1;
@@ -194,7 +194,7 @@ public class ExtDataGenerator extends DataGenerator {
 	 * @throws Exception
 	 */
 	private void executeAutoFilling(DataRecord[] outRecord) throws Exception {
-		for (int i = 0; i < recordsNumber && runIt; i++) {
+		for (long i = 0; i < recordsNumber && runIt; i++) {
 			for (int outPort = 0; outPort < outRecord.length; outPort++) {
 				autoFilling.setLastUsedAutoFillingFields(outRecord[outPort]);
 				writeRecord(outPort, outRecord[outPort]);
